workspaces/requirements.txt
python==3.11.5
ipykernel==6.25.2
pytorch==2.0.1
transformers==4.33.2
bertviz==1.4.0
ipywidgets==8.1.1